requirements.txt
asgiref==3.2.10
Django==3.0.7
djangorestframework==3.11.0
Faker==4.1.1
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
